change an int to ll

+++ tr_fast.cpp
@@ -55,11 +55,11 @@ void build_table(ll n, ll k) {
     ll group_sz_first = pow(k, c - 1) + EPS;
     ll repeats = 1;
     for (ll group_sz = group_sz_first; group_sz > 0; group_sz = group_sz / k) {
-        for (int r = 0; r < repeats; ++r) { // repeats of groups
+        for (ll r = 0; r < repeats; ++r) { // repeats of groups
             for (int c = 0; c < k; ++c) {
                 char to_append = c + CHAR_OFFSET;
 
-                for (int i = 0; i < group_sz; ++i) {
+                for (ll i = 0; i < group_sz; ++i) {
                     table[r * group_sz * k + c * group_sz + i] += to_append; 
                 }
             }
